refactor(logger): consolidate Done and Start methods into logger.go
