+++ logger/logger.go
@@ -1,7 +1,9 @@
 package logger
 
 import (
+	"fmt"
 	"io"
+	"time"
 )
 
 type Logger struct {
@@ -11,4 +13,18 @@ type Logger struct {
 // NewLogger creates a new Logger instance with the specified writer
 func NewLogger(output io.Writer) *Logger {
 	return &Logger{Output: output}
+}
+
+// Done logs the completion of a download operation.
+// It outputs the final URL and the timestamp at which the download finished,
+// formatted as "yyyy-mm-dd hh:mm:ss" for consistency with wget-style logging.
+func (l *Logger) Done(timestamp time.Time, url string) {
+	fmt.Fprintf(l.Output, "\nDownloaded [%s]\n", url)
+	fmt.Fprintf(l.Output, "finished at %s\n", timestamp.Format("2006-01-02 15:04:05"))
+}
+
+// Start logs the start of a process with a timestamp.
+func (l *Logger) Start(url string, timestamp time.Time) {
+	fmt.Fprintf(l.Output, "start at %s\n", timestamp.Format("2006-01-02 15:04:05"))
+	fmt.Fprintf(l.Output, "sending request, awaiting response... ")
 }
